+++ lib/simmer/suite/externals.rb
