+++ lib/simmer/suite/specification/assert/assertions.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+#
+# Copyright (c) 2020-present, Blue Marble Payroll, LLC
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+#
+
+require_relative 'assertions/output'
+require_relative 'assertions/table'
+
+module Simmer
+  class Suite
+    class Specification
+      class Assert
+        class Assertions
+          acts_as_hashable_factory
+
+          register 'output', Output
+          register 'table', Table
+        end
+      end
+    end
+  end
+end
